refactor: move static catalogs to data modules
